documentation, tweak to new shell, fixed NO_SCRIPTS bug
Added the files I use to make my documentation files, in case anyone needs to pick up where I leave off one day.

Added a bit in the new version of the Hugo shell (the one that uses a flags.h file) so USE_ROODYLIB is on by default

Remembered that when I improved the character script code last year, I forgot to check if my code worked with games where NO_SCRIPTS was set.  It didn't, and now it is fixed.